Add more methods and libs for qa framework
